moved user DTOs to its own package
