+++ api/src/main/java/tennisclub/dto/user/UserAuthDTO.java
@@ -1,4 +1,4 @@
-package tennisclub.dto;
+package tennisclub.dto.user;
 
 import org.hibernate.validator.constraints.Length;
 import javax.validation.constraints.NotBlank;
